// src/types/podcast_index.ts
interface PIApiFeedBase {
  /** The internal podcastindex.org feed id. */
  id: number;
  /** The feed title. */
  title: string;
  /** The current feed url. */
  url: string;
  /** The iTunes id of this feed if there is one, and we know what it is. */
  itunesId: number | null;
  /** The channel-level language specification of the feed. Languages accord with the RSS language spec. */
  language: string;
  /** category id: name, NOTE: this is not always present on the raw responses and will be populated with an empty object when missing form the api response */
  categories: {
    [k: string]: string;
  } | null;
}

/** Returned by search */
export interface PIApiFeed extends PIApiFeedBase {
  locked?: number;
  imageUrlHash?: number;
  /** The url of the feed, before it changed to it’s current url. */
  originalUrl: string;
  /** The channel level link in the feed. */
  link: string;
  /** The channel-level description. */
  description: string;
  /** The channel-level author element. Usually iTunes specific, but could be from another namespace if not present. */
  author: string;
  /** The channel-level owner:name element. Usually iTunes specific, but could be from another namespace if not present. */
  ownerName: string;
  /** The channel-level image element. */
  image: string;
  /** The seemingly best artwork we can find for the feed. Might be the same as ‘image’ in most instances. */
  artwork: string;
  /** [Unix Epoch] The channel-level pubDate for the feed, if it’s sane. If not, this is a heuristic value, arrived at by analyzing other parts of the feed, like item-level pubDates. */
  lastUpdateTime: number;
  /** [Unix Epoch] The last time we attempted to pull this feed from it’s url. */
  lastCrawlTime: number;
  /** [Unix Epoch] The last time we tried to parse the downloaded feed content. */
  lastParseTime: number;
  /** [Unix Epoch] Timestamp of the last time we got a "good", meaning non-4xx/non-5xx, status code when pulling this feed from it’s url. */
  lastGoodHttpStatusTime: number;
  /** The last http status code we got when pulling this feed from it’s url. You will see some made up status codes sometimes. These are what we use to track state within the feed puller. These all start with 9xx. */
  lastHttpStatus: number;
  /** The Content-Type header from the last time we pulled this feed from it’s url. */
  contentType: string;
  /** The channel-level generator element if there is one. */
  generator: string | null;
  /** 0 = RSS, 1 = ATOM */
  type: PodcastFeedType;
  /** At some point, we give up trying to process a feed and mark it as dead. This is usually after 1000 errors without a successful pull/parse cycle. Once the feed is marked dead, we only check it once per month. */
  dead: number;
  /** The number of errors we’ve encountered trying to pull a copy of the feed. Errors are things like a 500 or 404 response, a server timeout, bad encoding, etc. */
  crawlErrors: number;
  /** The number of errors we’ve encountered trying to parse the feed content. Errors here are things like not well-formed xml, bad character encoding, etc. We fix many of these types of issues on the fly when parsing. We only increment the errors count when we can’t fix it. */
  parseErrors: number;
}
// ...
export enum PodcastFeedType {
  RSS = 0,
  ATOM = 1,
}